move: 비회원 예약 조회 페이지의 잘못된 폴더 이름 수정
